+++ src/actions/profile.js
@@ -179,19 +179,21 @@ export const addEducation = (formData, history) => async dispatch => {
 }
 //Delete Education
 export const deleteEducation = id => async dispatch => {
-    try {
-        const res = await axios.delete(`http://localhost:5000/profile/education/${id}`);
-        dispatch({
-            type: UPDATE_PROFILE,
-            payload: res.data
-        })
-        dispatch(setAlert('Education Removed', 'warning'));
-    } catch (error) {
-        dispatch({
-            type: PROFILE_ERROR,
-            payload: { msg: error.response.statusText, status: error.response.status }
+    if (window.confirm('Are you sure? this can not be undone')) {
+        try {
+            const res = await axios.delete(`http://localhost:5000/profile/education/${id}`);
+            dispatch({
+                type: UPDATE_PROFILE,
+                payload: res.data
+            })
+            dispatch(setAlert('Education Removed', 'warning'));
+        } catch (error) {
+            dispatch({
+                type: PROFILE_ERROR,
+                payload: { msg: error.response.statusText, status: error.response.status }
 
-        });
+            });
+        }
     }
 }
 //Delete Account & profile
